# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverRead": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "auth_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Id"
          },
          "availability": {
            "items": {
              "type": "boolean"
            },
            "title": "Availability",
            "type": "array"
          },
          "car_make_model": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Car Make Model",
            "type": "string"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "readOnly": true,
            "title": "Full Name",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "license_plate": {
            "maxLength": 20,
            "minLength": 1,
            "title": "License Plate",
            "type": "string"
          },
          "note_chain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Chain Id"
          },
          "partner_driver_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Driver Name"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "license_plate",
          "car_make_model",
          "address",
          "driver_id",
          "user_id",
          "auth_id",
          "first_name",
          "last_name",
          "email",
          "role",
          "full_name"
        ],
        "title": "DriverRead",
        "type": "object"
      },
...
      "DriverRegisterResponse": {
        "description": "Driver registration response - contains Driver object and AuthResponse",
        "properties": {
          "auth": {
            "$ref": "#/components/schemas/AuthResponse"
          },
          "driver": {
            "$ref": "#/components/schemas/DriverRead"
          }
        },
        "required": [
          "driver",
          "auth"
        ],
        "title": "DriverRegisterResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserFinalize": {
        "properties": {
          "password": {
            "maxLength": 100,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "user_invite_id": {
            "format": "uuid",
            "title": "User Invite Id",
            "type": "string"
          }
        },
        "required": [
          "user_invite_id",
          "password"
        ],
        "title": "UserFinalize",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the Food4Kids application",
    "title": "Food4Kids API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/drivers/": {
      "get": {
        "description": "Get all drivers, optionally filter by driver_id or email",
        "operationId": "get_drivers",
        "parameters": [
          {
            "description": "Filter by driver ID",
            "in": "query",
            "name": "driver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by driver ID",
              "title": "Driver Id"
            }
          },
          {
            "description": "Filter by email",
            "in": "query",
            "name": "email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by email",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverRead"
                  },
                  "title": "Response Get Drivers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Drivers",
        "tags": [
          "drivers"
        ]
      }
    },
    "/drivers/initialize": {
      "post": {
        "description": "Register a new driver in our backend, creates a User and Driver object, returns DriverRead\nNOTE: This does not create a firebase user, ie the User is in a hanging state\nWe need to do this so that we can implement our invite only system",
        "operationId": "initialize_driver",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Initialize Driver",
        "tags": [
          "drivers"
        ]
      }
    },
    "/drivers/register": {
      "post": {
        "description": "Creates Firebase user and attaches to hanging state user in our local db, returns DriverRegisterResponse",
        "operationId": "complete_driver_registration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFinalize"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Driver Registration",
        "tags": [
          "drivers"
        ]
      }
    },
    "/drivers/test-event-email": {
      "post": {
        "description": "Temporary endpoint to test event-driven emails.\nDelete this after testing!",
        "operationId": "test_event_email",
        "parameters": [
          {
            "in": "query",
            "name": "test_email",
            "required": true,
            "schema": {
              "title": "Test Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Test Event Email",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Event Email",
        "tags": [
          "drivers"
        ]
      }
    },
    "/drivers/{driver_id}": {
      "delete": {
        "description": "Delete a driver by ID.\n\nTheir routes and mileage adjustments are detached (driver_id SET NULL),\nso the driver's km stop counting toward anyone.",
        "operationId": "delete_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Driver",
        "tags": [
          "drivers"
        ]
      },
      "get": {
        "description": "Get a single driver by ID",
        "operationId": "get_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver",
        "tags": [
          "drivers"
        ]
      },
      "put": {
        "description": "Update an existing driver",
        "operationId": "update_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Driver",
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/routes": {
      "get": {
        "description": "Get routes with pagination and optional filtering for unassigned routes and date range.\nReturns routes with their drive dates - routes can appear multiple times for different dates.\nWhen unassigned_only is False, returns all routes (no assignment filter).\nWhen unassigned_only is True, returns only routes that are unassigned for the given route group.\n\nRequires a driver or admin caller.\nAdmins may scope to any driver via driver_id (or omit it for all routes).\nDrivers are always scoped to their own routes: omitting driver_id returns\ntheir own routes, and requesting another driver's is rejected.",
        "operationId": "get_routes",
        "parameters": [
          {
            "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
            "in": "query",
            "name": "unassigned_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
              "title": "Unassigned Only",
              "type": "boolean"
            }
          },
          {
            "description": "Filter route groups from this date",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "description": "Filter route groups from this date",
              "title": "Start Date",
              "type": "string"
            }
          },
          {
            "description": "Filter route groups until this date",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "description": "Filter route groups until this date",
              "title": "End Date",
              "type": "string"
            }
          },
          {
            "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Number of items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
            "in": "query",
            "name": "driver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RouteWithDateRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Routes",
        "tags": [
          "routes"
        ]
      }
    },
    "/routes/{route_id}": {
      "delete": {
        "description": "Delete a route by its unique identifier.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to delete.\n    session (AsyncSession): The database session dependency.\n    _ (bool): Indicates that the user is authenticated as a driver (injected by dependency).\n\nReturns:\n    None. Responds with HTTP 204 No Content on successful deletion.",
        "operationId": "delete_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Route",
        "tags": [
          "routes"
        ]
      },
      "get": {
        "description": "Get a route by its unique identifier, with its ordered stops embedded.\n\nEach stop carries its sequence #, address, contact name, phone (+ secondary),\nbox count, and a note_chain_id reference. Stops are sourced live from\nLocation for upcoming routes and from frozen route_stop_snapshots for past\nroutes. Notes are not embedded: fetch them via GET /note-chains/{id}/notes.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to GET.\n    session (AsyncSession): The database session dependency.\n\nReturns:\n    The route with its ordered stops. Responds with HTTP 200 OK on success.",
        "operationId": "get_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route",
        "tags": [
          "routes"
        ]
      },
      "patch": {
        "description": "Update a route's metadata (name, notes) and/or its stop order/locations.\n\nIf location_ids is provided in the request body:\n- The route's stops are fully replaced with the new ordered list.\n- The routing algorithm is re-run to compute the new polyline and mileage.\n- This will affect ALL route groups that share this route.\n\nIf only name/notes are provided, stops and mileage are left unchanged.",
        "operationId": "update_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Route",
        "tags": [
          "routes"
        ]
      }
    },
...
  }
}